// utils.d.ts
export declare const divisors: {
    days: number;
    hours: number;
    minutes: number;
    seconds: number;
};
export declare const endOfMonth: (date: Date) => Date;
export declare const endOfYear: (date: Date) => Date;
export declare const beginOfYear: (date: Date) => Date;
export declare const dayOfYear: (date: Date) => number;
export declare const daysInYear: (date: Date) => number;
export declare const roundToPrecision: (n: number, decimalPrecision: number) => number;